fix: edited some ewram vars to be in line with the style guide

+++ include/pokedex_area_screen.h
@@ -3,7 +3,7 @@
 
 #include "rtc.h"
 
-extern u32 gAreaTimeOfDay;
+extern u8 gAreaTimeOfDay;
 
 enum PokedexAreaScreenState
 {
